+++ packages/cloudflare-access/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/cloudflare-access [1.0.0-alpha.46](https://github.com/anolilab/lunora/compare/%40lunora%2Fcloudflare-access%401.0.0-alpha.45...%40lunora%2Fcloudflare-access%401.0.0-alpha.46) (2026-07-27)
+
+
+### Dependencies
+
+* **@lunora/server:** upgraded to 1.0.0-alpha.48
+
 ## @lunora/cloudflare-access [1.0.0-alpha.45](https://github.com/anolilab/lunora/compare/%40lunora%2Fcloudflare-access%401.0.0-alpha.44...%40lunora%2Fcloudflare-access%401.0.0-alpha.45) (2026-07-27)